+++ app/src/main/java/com/techcourse/support/web/filter/DispatcherServletInitializer.java
@@ -1,9 +1,10 @@
-package com.techcourse;
+package com.techcourse.support.web.filter;
 
 import jakarta.servlet.ServletContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import web.org.springframework.web.WebApplicationInitializer;
+import webmvc.org.springframework.web.servlet.mvc.DispatcherServlet;
 
 /**
  * Base class for {@link WebApplicationInitializer}
