+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/DispatcherServlet.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package webmvc.org.springframework.web.servlet.mvc;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
@@ -8,6 +8,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import webmvc.org.springframework.web.servlet.ModelAndView;
 import webmvc.org.springframework.web.servlet.View;
+import webmvc.org.springframework.web.servlet.mvc.exception.AdapterNotFoundException;
+import webmvc.org.springframework.web.servlet.mvc.exception.HandlerNotFoundException;
 import webmvc.org.springframework.web.servlet.mvc.tobe.AnnotationHandlerMapping;
 import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerAdapter;
 import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerMapping;
@@ -43,11 +45,11 @@ public class DispatcherServlet extends HttpServlet {
         final Object handler = handlerMappings.stream()
                 .map(mapping -> mapping.getHandler(request))
                 .findFirst()
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않은 요정입니다."));
+                .orElseThrow(() -> new HandlerNotFoundException("존재하지 않는 요청입니다."));
         final HandlerAdapter handlerAdapter = handlerAdapters.stream()
                 .filter(Adapter -> Adapter.supports(handler))
                 .findFirst()
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않은 요정입니다."));
+                .orElseThrow(() -> new AdapterNotFoundException("정의하지 않는 요청입니다."));
 
         try {
             final ModelAndView handle = handlerAdapter.handle(request, response, handler);
