+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/exception/AdapterNotFoundException.java
@@ -1,0 +1,7 @@
+package webmvc.org.springframework.web.servlet.mvc.exception;
+
+public class AdapterNotFoundException extends RuntimeException {
+    public AdapterNotFoundException(String message) {
+        super(message);
+    }
+}
