+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/exception/HandlerNotFoundException.java
@@ -1,0 +1,7 @@
+package webmvc.org.springframework.web.servlet.mvc.exception;
+
+public class HandlerNotFoundException extends RuntimeException {
+    public HandlerNotFoundException(String message) {
+        super(message);
+    }
+}
